Agregando movimiento de pezcado variado, cambiando velocidad de juego, pezcado y stalkers

+++ Fish-Game/src/fish/game/ObstacleM.java
@@ -33,6 +33,6 @@ public class ObstacleM extends Item{
 
     @Override
     public void render(Graphics g) {
-        g.drawImage(Assets.flores,getX(),getY(),getWidth(),getHeight(),null);
+        g.drawImage(Assets.flores,(int) getX(),(int) getY(),getWidth(),getHeight(),null);
     }
 }
